Add files via upload
Initial release of STM8S003F3 based modbus ASCII/RTU ligth switch module.
This is my hobby project of home automatization system based on modbus protocol. This module is projected to be the cheapest of I saw in the market.

Supported:
- Modbus ASCII/RTU protocol. Baud rates: 4800, 9600,19200,38400,57600,115200
- IR receiver of NEC compatable protocol remote control. 32 to 96 bits of data and different preambles are supported
- IR transmitter at 38KHz (depends on Timer2 configuration) is supported. NEC compatable protocol 32-96bits are supported.
- DS18B20 is supported, it is working, but.. I can't disable interrupts that is why it will read temperature not every time it try to do it.
- Read 3 GPIOs with anti-glitch via modbus are supported.
Write 3 GPIOs via modbus are supported. It is developed to light buttons.

You can use the project and source code at your discretion. And as they usually say: "The developer is not responsible for any damage, directly or indirectly."